azurerm/internal/services/eventhub/parse/eventhub_cluster.go
package parse

import (
	"fmt"

	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
)

type EventHubClusterId struct {
	ResourceGroup string
	Name          string
}

func EventHubClusterID(input string) (*EventHubClusterId, error) {
	id, err := azure.ParseAzureResourceID(input)
	if err != nil {
		return nil, fmt.Errorf("parsing EventHub Cluster ID %q: %+v", input, err)
	}

	cluster := EventHubClusterId{
		ResourceGroup: id.ResourceGroup,
	}

	if cluster.Name, err = id.PopSegment("clusters"); err != nil {
		return nil, err
	}

	if err := id.ValidateNoEmptySegments(input); err != nil {
		return nil, err
	}

	return &cluster, nil
}
